backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
Django==5.2.10
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-filter==25.2
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==24.1.1
idna==3.11
kombu==5.6.2
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
stripe==14.2.0
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
